+++ src/DuetAPI/Commands/Generic/Code/Parser.cs
@@ -34,7 +34,7 @@ namespace DuetAPI.Commands
             bool contentRead = false, unprecedentedParameter = false;
             bool inFinalComment = false, inEncapsulatedComment = false, inChunk = false, inQuotes = false, inExpression = false, inCondition = false;
             bool readingAtStart = true, isLineNumber = false, isNumericParameter = false, endingChunk = false;
-            bool wasQuoted = false, wasExpression = false;
+            bool nextCharLowerCase = false, wasQuoted = false, wasExpression = false;
             int numCurlyBraces = 0, numRoundBraces = 0;
 
             char[] charArray = new char[1];
@@ -144,7 +144,20 @@ namespace DuetAPI.Commands
                 {
                     if (inQuotes)
                     {
-                        if (c == '"')
+                        if (c == '\'')
+                        {
+                            if (nextCharLowerCase)
+                            {
+                                // Treat subsequent single-quotes as a single-quite char
+                                value += '\'';
+                            }
+                            else
+                            {
+                                // Next letter should be lower-case
+                                nextCharLowerCase = true;
+                            }
+                        }
+                        else  if (c == '"')
                         {
                             if (reader.Peek() == '"')
                             {
@@ -156,10 +169,16 @@ namespace DuetAPI.Commands
                             else
                             {
                                 // No longer in an escaped parameter
-                                inQuotes = false;
+                                inQuotes = nextCharLowerCase = false;
                                 wasQuoted = true;
                                 endingChunk = true;
                             }
+                        }
+                        else if (nextCharLowerCase)
+                        {
+                            // Add next lower-case character to the parameter value
+                            value += char.ToLower(c);
+                            nextCharLowerCase = false;
                         }
                         else
                         {